refactor(gantz): extract persistence into a persist module
Move the Persister, the background worker, and the persist/exit-flush systems
out of main.rs into a self-contained persist::PersistPlugin. main.rs now just
adds the plugin, leaving the store + the shared debounce in the app. A step
toward lifting this into a reusable sync-friendly-persistence crate so
downstream gantz apps don't reimplement the boilerplate.

Also expands the Pkv doc to explain why the Arc<Mutex> never contends during
the steady-state render loop: the main thread only reads the store, all reads
happen at startup / first frame before any persist can fire, and runtime writes
go to the worker over a channel rather than locking the store.
