+++ crates/gantz/src/storage.rs
@@ -11,10 +11,17 @@ use std::sync::{Arc, Mutex};
 /// A [`Resource`] wrapping a shared [`PkvStore`] that implements [`Load`] and
 /// [`Save`].
 ///
-/// `Arc<Mutex<_>>` so a background persistence worker can share the single store
-/// handle with the main thread (which reads at startup and on the first frame).
-/// redb permits only one handle per database file, so sharing - rather than a
-/// second handle - is required.
+/// `Arc<Mutex<_>>` lets a background persistence worker share the single store
+/// handle with the main thread - redb permits only one handle per database file,
+/// so sharing, rather than a second handle, is required.
+///
+/// Despite the `Mutex`, there is no lock contention during the steady-state
+/// render loop. The main thread only ever *reads* the store, and every read
+/// happens during startup and the first-frame egui-memory load - before any
+/// debounced persist can fire (a persist needs prior input). From then on the
+/// worker is the sole accessor: the persist systems hand their writes to it over
+/// a channel rather than locking the store themselves. So the worker's fsync'd
+/// writes, however long they take, never block a frame.
 #[derive(Resource, Clone)]
 pub struct Pkv(pub Arc<Mutex<PkvStore>>);
